requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.1
langchain-core>=0.1.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0
jinja2>=3.0.0
pyyaml>=6.0.0

# Development
rich>=13.0.0
pytest>=8.0.0
